+++ modules/deployment-model/src/main/java/org/mule/runtime/deployment/model/api/plugin/ArtifactPluginClassLoaderFactory.java
@@ -10,8 +10,8 @@ package org.mule.runtime.deployment.model.api.plugin;
 import org.mule.runtime.module.artifact.api.classloader.ArtifactClassLoader;
 import org.mule.runtime.module.artifact.api.classloader.ArtifactClassLoaderFactory;
 import org.mule.runtime.module.artifact.api.classloader.ClassLoaderLookupPolicy;
-import org.mule.runtime.module.artifact.api.classloader.MuleArtifactClassLoader;
 import org.mule.runtime.module.artifact.api.descriptor.ArtifactPluginDescriptor;
+import org.mule.runtime.module.artifact.internal.classloader.MulePluginClassLoader;
 
 /**
  * Creates {@link ArtifactClassLoader} for application or domain plugin descriptors.
@@ -23,8 +23,8 @@ public final class ArtifactPluginClassLoaderFactory implements ArtifactClassLoad
   public ArtifactClassLoader create(String artifactId, ArtifactPluginDescriptor descriptor,
                                     ClassLoader parent,
                                     ClassLoaderLookupPolicy lookupPolicy) {
-    return new MuleArtifactClassLoader(artifactId, descriptor, descriptor.getClassLoaderConfiguration().getUrls(), parent,
-                                       lookupPolicy);
+    return new MulePluginClassLoader(artifactId, descriptor, descriptor.getClassLoaderConfiguration().getUrls(), parent,
+                                     lookupPolicy);
   }
