+++ modules/service/src/main/java/org/mule/runtime/module/service/api/manager/ServiceProxyInvocationHandler.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) MuleSoft, Inc.  All rights reserved.  http://www.mulesoft.com
+ * The software in this package is published under the terms of the CPAL v1.0
+ * license, a copy of which has been included with this distribution in the
+ * LICENSE.txt file.
+ */
+package org.mule.runtime.module.service.api.manager;
+
+import org.mule.runtime.api.service.Service;
+
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Proxy;
+
+/**
+ * Provides access to the actual service wrapped within a {@link Proxy}.
+ * 
+ * @since 4.5
+ */
+public interface ServiceProxyInvocationHandler extends InvocationHandler {
+
+  /**
+   * @return the proxied service. May be {@code null} before the first proxy is started.
+   */
+  public abstract Service getService();
+
+}
