+++ modules/service/src/main/java/org/mule/runtime/module/service/internal/manager/ServiceRegistry.java
@@ -8,8 +8,10 @@ package org.mule.runtime.module.service.internal.manager;
 
 import static java.lang.String.format;
 import static java.util.Optional.ofNullable;
+
 import static org.reflections.ReflectionUtils.getAllFields;
 import static org.reflections.ReflectionUtils.withAnnotation;
+
 import org.mule.runtime.api.service.Service;
 import org.mule.runtime.api.service.ServiceProvider;
 import org.mule.runtime.module.service.api.discoverer.ServiceAssembly;
@@ -18,6 +20,7 @@ import org.mule.runtime.module.service.api.discoverer.ServiceResolutionError;
 import java.lang.reflect.Field;
 import java.lang.reflect.ParameterizedType;
 import java.lang.reflect.Type;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -66,7 +69,8 @@ public class ServiceRegistry {
           field.set(serviceProvider, dependency);
         } else if (!asOptional) {
           throw new ServiceResolutionError(format("Cannot find a service to inject into field '%s' of service provider '%s'",
-                                                  field.getName(), dependencyType.getName()));
+                                                  field.getName(),
+                                                  serviceProvider.getServiceDefinition().getServiceClass().getName()));
         }
       } catch (Exception e) {
         throw new ServiceResolutionError(format("Could not inject dependency on field %s of type %s", field.getName(),
@@ -96,4 +100,24 @@ public class ServiceRegistry {
     services.put(assembly.getServiceContract(), service);
   }
 
+  public <S extends Service> void register(Class<? extends S> serviceContract, S service) {
+    services.put(serviceContract, service);
+  }
+
+  public <S extends Service> void unregister(Class<? extends S> serviceContract) {
+    services.remove(serviceContract);
+  }
+
+  public <S extends Service> Optional<S> getService(Class<S> serviceInterface) {
+    return services
+        .values()
+        .stream()
+        .filter(s -> serviceInterface.isInstance(s))
+        .map(s -> (S) s)
+        .findAny();
+  }
+
+  public Collection<Service> getAllServices() {
+    return services.values();
+  }
 }
